Require abort on malformed extension
This is to address a comment by Michael StJohns on the mailing list. To quote:
```
Section 2 requires  (MUST) the generation of fatal illegal_parameter alert upon reception of a mal-encoded 
extension (e.g. any trailing zero bytes), but compare and contrast this with section 3 which is full of MUST 
and MUST NOT declarations but with no concrete actions to be taken.  E.g. if I (server or client) send 0x01 0x10, 
and receive 0x01 0x11 from the client or server, wouldn't that be an illegal value as I've added a bit not sent 
to me?   Should that cause the same fatal illegal_parameter alert? Alternately, "receiver MUST ignore received 
bits that weren't sent" language could clean this up.
```