Clarify language around the size of the extension.
The declarative syntax specifies that flags are <1..255>, and
therefore can not be empty. If it does not conform to the
syntax, we should fail with a decode error alert.

The MUST NOT implies that this is illegal for some other reason
and therefore would require failing with an illegal parameter alert.

Since the MUST NOT is redundant with the declarative syntax
as specified, just remove this.